requirements.txt
SQLAlchemy==1.3.20
paho-mqtt~=1.5.1